requirements.txt
pip~=22.3.1
Pillow~=9.4.0
Django~=4.2.3
asgiref~=3.7.2
sqlparse~=0.4.4
colorama~=0.4.6
setuptools~=68.0.0
psycopg2-binary
clients~=1.4
APScheduler~=3.10.4
pathlib~=1.0.1
defusedxml~=0.7.1
cffi~=1.15.1
future~=0.18.3
ipython~=8.15.0
zlib~=1.2.13
jupyter~=1.0.0
panel~=1.2.3
attrs~=22.1.0
wheel~=0.38.4
openssl~=3.0.11
cryptography~=41.0.3
docutils~=0.18.1
sphinx~=5.0.2
tornado~=6.3.2
jinja2~=3.1.2
parso~=0.8.3
jedi~=0.18.1
pytz~=2023.3.post1
yaml~=0.2.5
pyyaml~=6.0
tblib~=1.7.0
numpy~=1.24.3
requests~=2.31.0
click~=8.0.4
typing_extensions~=4.7.1
pygments~=2.15.1
pytest~=7.4.0
traitlets~=5.7.1
stack_data~=0.2.0
decorator~=5.1.1
matplotlib~=3.7.2
packaging~=23.1
backcall~=0.2.0
pickleshare~=0.7.5
jupyter_client~=7.4.9
pexpect~=4.8.0
prompt_toolkit~=3.0.36
wcwidth~=0.2.5
six~=1.16.0
asttokens~=2.0.5
astroid~=2.14.2
executing~=0.8.3
pure_eval~=0.2.2
djangorestframework~=3.14.0
markdown~=3.4.1